Switch to using XCTest for specs
All tests have been converted from Quick/Nimble over to using the out of
the box XCTest. Quick and Nimble were great, but I wanted to reduce the
barrier to entry for contributors by using the simplest testing
framework.

In addition, this will make it easier for me to get tests working in
pre-release versions of Swift and Xcode.
