--- Cartfile.private
@@ -1,2 +1,0 @@
-github "Quick/Quick"
-github "Quick/Nimble"
