--- Cartfile.resolved
@@ -1,2 +1,0 @@
-github "Quick/Nimble" "v3.2.0"
-github "Quick/Quick" "v0.9.1"
